pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saa"
version = "0.1.0"
description = "Site Audit Agent - CLI tool for automated website audits"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.1.0",
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
saa = "saa.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/saa"]

[tool.hatch.build.targets.sdist]
include = ["src/saa"]
